move top mention model to the common module
